[IMP] trevly: modify template for sale.order and account.move reports
The client provided a series of features that want to be changed into
the actual template of a sale order and an invoice PDF. These ones
improves the amount of data shown in less space, more lines covered into
the same page, styles and more.
Changes into related reports, new paper format, and own new layout were
used to cover the needs.