[FIX] trevly: add display_type restriction in write of purchase lines
In 16.0, It was merged a solution from [1] for displayed lines, but,
currently, it is not available to do a backport to 15.0 for Odoo and
this
is for odoo.sh, so the patches are not available. Therefore, the write
is inherited to remove the key "product_qty" when a displayed line will
be written.

References:
- [1] https://github.com/odoo/odoo/pull/168676